+++ include/blmc_robots/n_joint_blmc_robot_driver.hpp
@@ -1,8 +1,6 @@
 /**
- * \file test_bench_8_motors.hh
- * \brief The hardware wrapper of the real Finger robot.
- * \author Manuel Wuthrich
- * \date 2018
+ * \file
+ * \brief Base driver for a generic n-joint BLMC robot.
  * \copyright Copyright (c) 2019, New York University and Max Planck
  *            Gesellschaft.
  */
@@ -12,12 +10,8 @@
 #include <math.h>
 #include <Eigen/Eigen>
 #include <blmc_robots/common_header.hpp>
-#include <tuple>
 
 #include <blmc_robots/blmc_joint_module.hpp>
-#include <real_time_tools/spinner.hpp>
-#include <real_time_tools/timer.hpp>
-
 #include <robot_interfaces/finger.hpp>
 
 namespace blmc_robots
@@ -63,7 +57,7 @@ struct CalibrationParameters
  * @brief Base class for simple n-joint BLMC robots.
  *
  * This is a generic base class to easily implement drivers for simple BLMC
- * robots that consist of a chain of joints.
+ * robots that consist of N_JOINTS joints.
  *
  * @tparam N_JOINTS Number of joints.
  * @tparam N_MOTOR_BOARDS Number of motor control boards that are used.
@@ -390,163 +384,5 @@ public:
     }
 };
 
-class RealFingerDriver : public NJointBlmcRobotDriver<3, 2>
-{
-public:
-    RealFingerDriver(const std::string &can_0, const std::string &can_1)
-        : RealFingerDriver(create_motor_boards({can_0, can_1}))
-    {
-    }
-
-private:
-    RealFingerDriver(const MotorBoards &motor_boards)
-        : NJointBlmcRobotDriver<3, 2>(motor_boards,
-                                      create_motors(motor_boards),
-                                      { // MotorParameters
-                                          .max_current_A = 2.0,
-                                          .torque_constant_NmpA = 0.02,
-                                          .gear_ratio = 9.0,
-                                      },
-                                      { // CalibrationParameters
-                                          .torque_ratio = 0.6,
-                                          .control_gain_kp = 3.0,
-                                          .control_gain_kd = 0.03,
-                                          .position_tolerance_rad = 0.05,
-                                          .move_timeout = 2000,
-                                      },
-                                      Vector(0.08, 0.08, 0.04))
-    {
-        home_offset_rad_ << -0.54, -0.17, 0.0;
-        initial_position_rad_ << 1.5, 1.5, 3.0;
-    }
-
-    static Motors create_motors(const MotorBoards &motor_boards)
-    {
-        // set up motors
-        Motors motors;
-        motors[0] = std::make_shared<blmc_drivers::Motor>(motor_boards[0], 0);
-        motors[1] = std::make_shared<blmc_drivers::Motor>(motor_boards[0], 1);
-        motors[2] = std::make_shared<blmc_drivers::Motor>(motor_boards[1], 0);
-
-        return motors;
-    }
-};
-
-robot_interfaces::FingerTypes::BackendPtr create_real_finger_backend(
-    const std::string &can_0,
-    const std::string &can_1,
-    robot_interfaces::FingerTypes::DataPtr robot_data)
-{
-    constexpr double MAX_ACTION_DURATION_S = 0.003;
-    constexpr double MAX_INTER_ACTION_DURATION_S = 0.005;
-
-    std::shared_ptr<robot_interfaces::RobotDriver<
-        robot_interfaces::FingerTypes::Action,
-        robot_interfaces::FingerTypes::Observation>>
-        robot = std::make_shared<RealFingerDriver>(can_0, can_1);
-
-    auto backend = std::make_shared<robot_interfaces::FingerTypes::Backend>(
-        robot, robot_data, MAX_ACTION_DURATION_S, MAX_INTER_ACTION_DURATION_S);
-    backend->set_max_action_repetitions(-1);
-
-    return backend;
-}
-
-
-// FIXME move this out of here!
-
-/**
- * @brief Driver for a single joint.
- *
- * Driver for a single BLMC joint.  Mostly intended for testing purposes.
- */
-class OneJointDriver : public NJointBlmcRobotDriver<1, 1>
-{
-public:
-
-    /**
-     * @brief Constructor
-     *
-     * @param can_port  Name of the CAN port to which the joint is connected
-     *     (e.g. "can0")
-     * @param home_offset_rad  Home offset of the joint.  This is the offset
-     *     between the home position (= encoder index found during homing) and
-     *     the desired zero position.
-     */
-    OneJointDriver(const std::string &can_port,
-                   const double home_offset_rad=0.0)
-        : OneJointDriver(create_motor_boards({can_port}), home_offset_rad)
-    {
-    }
-
-private:
-    OneJointDriver(const MotorBoards &motor_boards,
-                   const double home_offset_rad)
-        : NJointBlmcRobotDriver<1, 1>(motor_boards,
-                                      create_motors(motor_boards),
-                                      { // MotorParameters
-                                          .max_current_A = 2.0,
-                                          .torque_constant_NmpA = 0.02,
-                                          .gear_ratio = 9.0,
-                                      },
-                                      { // CalibrationParameters
-                                          .torque_ratio = 0.6,
-                                          .control_gain_kp = 3.0,
-                                          .control_gain_kd = 0.03,
-                                          .position_tolerance_rad = 0.05,
-                                          .move_timeout = 2000,
-                                      },
-                                      make_vector(0.08))
-    {
-        home_offset_rad_ << home_offset_rad;
-        initial_position_rad_ << -home_offset_rad;
-    }
-
-    static Motors create_motors(const MotorBoards &motor_boards)
-    {
-        // set up motors
-        Motors motors;
-        motors[0] = std::make_shared<blmc_drivers::Motor>(motor_boards[0], 0);
-
-        return motors;
-    }
-
-    /**
-     * @brief Create a Vector with the given value.
-     *
-     * This is a helper function needed for initializing Vectors with a
-     * specific value upon creation.
-     *
-     * @param value Value to be set in the vector.
-     *
-     * @return The Vector.
-     */
-    static Vector make_vector(double value)
-    {
-        Vector vec;
-        vec << value;
-        return vec;
-    }
-};
-
-robot_interfaces::NJointRobotTypes<1>::BackendPtr create_one_joint_backend(
-    const std::string &can_0,
-    const double home_offset_rad,
-    robot_interfaces::NJointRobotTypes<1>::DataPtr robot_data)
-{
-    constexpr double MAX_ACTION_DURATION_S = 0.003;
-    constexpr double MAX_INTER_ACTION_DURATION_S = 0.005;
-
-    std::shared_ptr<robot_interfaces::RobotDriver<
-        robot_interfaces::NJointRobotTypes<1>::Action,
-        robot_interfaces::NJointRobotTypes<1>::Observation>>
-        robot = std::make_shared<OneJointDriver>(can_0, home_offset_rad);
-
-    auto backend = std::make_shared<robot_interfaces::NJointRobotTypes<1>::Backend>(
-        robot, robot_data, MAX_ACTION_DURATION_S, MAX_INTER_ACTION_DURATION_S);
-    backend->set_max_action_repetitions(-1);
-
-    return backend;
-}
-
 }  // namespace blmc_robots
+
